% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{append_times_ERP}
\alias{append_times_ERP}
\title{Append ERPs to voltage file.}
\usage{
append_times_ERP(voltage_dat, ERP, samp_freq_Hz, full_window_bounds)
}
\arguments{
\item{voltage_dat}{Formatted voltage dataframe.}

\item{ERP}{List of ERP components and their lower/upper bounds.}

\item{samp_freq_Hz}{Sampling frequency in Hz.}

\item{full_window_bounds}{Bounds of time window is ms e.g. c(-200, 1198).}
}
\value{
Voltage with ERPs appropriately appendended (dataframe).
}
\description{
Append ERPs to voltage file.
}
