% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{change_file_endings}
\alias{change_file_endings}
\title{Mass change file endings.}
\usage{
change_file_endings(path, ending_detect, ending_replace)
}
\arguments{
\item{path}{A number.}

\item{ending_detect}{File ending to change e.g. ".vmrk"}

\item{ending_replace}{File ending to change e.g. ".csv"}
}
\value{
Nothing in console, but files will be renamed.
}
\description{
Changes file types by editing endings.
}
\examples{
change_file_endings("project_files/", ".vmrk", ".csv")

}
