% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{ERP_to_times}
\alias{ERP_to_times}
\title{Convert ERP list to dataframe with times.}
\usage{
ERP_to_times(ERP_list, time_vec)
}
\arguments{
\item{ERP_list}{List of ERP components and their lower/upper bounds.}

\item{time_vec}{Vector of times.}
}
\value{
Dataframe with ERP and times.
}
\description{
Convert ERP list to dataframe with times.
}
