% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{get_markers}
\alias{get_markers}
\title{Format .vmrk marker file.}
\usage{
get_markers(filename, markers_exclude = NULL)
}
\arguments{
\item{filename}{Marker filename, full path.}

\item{markers_exclude}{Character vector of (response) markers to exclude e.g.
"S240" or if all >200 then paste0("S", 201:300)}
}
\value{
Formatted marker dataframe.
}
\description{
Format .vmrk marker file.
}
