% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{get_voltages}
\alias{get_voltages}
\title{Format .dat voltage file.}
\usage{
get_voltages(filename, var_labs, ERP, samp_freq_Hz, full_window_bounds)
}
\arguments{
\item{filename}{Voltage filename, full path.}

\item{var_labs}{Variable labels to assign to filename
parts separated by underscores e.g. c("subject", "condition").}

\item{ERP}{List of ERP components and their lower/upper bounds.}

\item{samp_freq_Hz}{Sampling frequency in Hz.}

\item{full_window_bounds}{Bounds of time window is ms e.g. c(-200, 1198).}
}
\value{
Formatted voltage dataframe.
}
\description{
Format .dat voltage file.
}
