% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{one_vhdr}
\alias{one_vhdr}
\title{BVA_to_R function but for only one file}
\usage{
one_vhdr(
  path,
  filename,
  var_labs,
  sep,
  ERP_list,
  full_window_bounds,
  markers_exclude = NULL
)
}
\arguments{
\item{path}{Full path to all .vhdr, .vmrk and .dat files.
Of the form "path/to/file/" with the final forward slash.
If unspecified, defaults to current working directory.}

\item{filename}{Vhdr filename, full path}

\item{var_labs}{Variable labels to assign to filename
parts separated by underscores e.g. c("subject", "condition").}

\item{sep}{Separation in the filename e.g. "_"}

\item{ERP_list}{Named list of ERP components and their lower/upper bounds.}

\item{full_window_bounds}{Bounds of time window is ms e.g. c(-200, 1198).}

\item{markers_exclude}{Character vector of (response) markers to exclude e.g.
"S240" or if all >200 then paste0("S", 201:300)}
}
\description{
BVA_to_R function but for only one file
}
