% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVA_to_R.R
\name{BVA_to_R}
\alias{BVA_to_R}
\title{Pipeline from BVA exports to R dataframe.}
\usage{
BVA_to_R(
  path,
  ERP,
  var_labs,
  samp_freq_Hz,
  full_window_bounds,
  filename_returned = NULL
)
}
\arguments{
\item{path}{Full path to all .vmrk and .dat files.}

\item{ERP}{List of ERP components and their lower/upper bounds.}

\item{var_labs}{Variable labels to assign to filename
parts separated by underscores e.g. c("subject", "condition").}

\item{samp_freq_Hz}{Sampling frequency in Hz.}

\item{full_window_bounds}{Bounds of time window is ms e.g. c(-200, 1198).}

\item{filename_returned}{"filename.csv" if you want a .csv returned/}
}
\value{
Formatted dataframe.
}
\description{
Take BVA exports and return large csv file
with eveyrthing summarised.
}
