% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{get_markers}
\alias{get_markers}
\title{Format .vmrk marker file.}
\usage{
get_markers(filename, var_labs)
}
\arguments{
\item{filename}{Marker filename, full path.}

\item{var_labs}{Variable labels to assign to filename
parts separated by underscores e.g. c("subject", "condition").}
}
\value{
Formatted marker dataframe.
}
\description{
Format .vmrk marker file.
}
